backend/requirements.txt
Flask
flask-cors
python-dotenv
pymongo
google-generativeai
